Move SES event handling into worker
